fix: scrolltoBookmark, addNotes not fix: showNotes

+++ app/src/main/java/raf/console/chitalka/data/local/room/NoteDao.kt
@@ -29,27 +29,27 @@ interface NoteDao {
     suspend fun delete(note: NoteEntity)
 
     @Query("DELETE FROM NoteEntity WHERE id IN (:ids)")
-    suspend fun deleteByIds(ids: List<Int>)
+    suspend fun deleteByIds(ids: List<Long>)
 
     @Query("SELECT * FROM NoteEntity ORDER BY createdAt DESC")
     suspend fun getAll(): List<NoteEntity>
 
     @Query("SELECT * FROM NoteEntity WHERE bookId = :bookId ORDER BY chapterIndex, offsetStart")
-    suspend fun getForBook(bookId: Int): List<NoteEntity>
+    suspend fun getForBook(bookId: Long): List<NoteEntity>
 
     @Query("SELECT * FROM NoteEntity WHERE bookmarkId = :bookmarkId")
-    suspend fun getForBookmark(bookmarkId: Int): List<NoteEntity>
+    suspend fun getForBookmark(bookmarkId: Long): List<NoteEntity>
 
     @Query("SELECT * FROM NoteEntity WHERE id = :id")
-    suspend fun getById(id: Int): NoteEntity?
+    suspend fun getById(id: Long): NoteEntity?
 
     @Query("SELECT COUNT(*) FROM NoteEntity")
-    suspend fun count(): Int
+    suspend fun count(): Long
 
     @Query("SELECT * FROM NoteEntity ORDER BY createdAt DESC")
     fun observeAll(): Flow<List<NoteEntity>>
 
     // 🔹 Новый метод для получения заметок по закладке через Flow
     @Query("SELECT * FROM NoteEntity WHERE bookmarkId = :bookmarkId")
-    fun observeForBookmark(bookmarkId: Int): Flow<List<NoteEntity>>
+    fun observeForBookmark(bookmarkId: Long): Flow<List<NoteEntity>>
 }
